Hyeryun you

+++ hyeryunYou/1주차/[11382] 꼬마 정민.c
@@ -2,8 +2,8 @@
 
 int main()
 {
-    int A, B, C;
-    scanf("%d %d %d", &A, &B, &C);
-    printf("%d", A+B+C);
+    long long A, B, C;
+    scanf("%lld %lld %lld", &A, &B, &C);
+    printf("%lld", A+B+C);
     return 0;
 }
